requirements.txt
asyncio
django
django-cors-headers
django-filter
djangorestframework
discord
DiscordUtils
psycopg2
python-dotenv
PyNaCl
python-dotenv
yt-dlp
requests
